# Conflicts:
#	composer.json
#	src/Viserio/Support/composer.json
#	src/Viserio/Translator/composer.json

+++ src/Viserio/Filesystem/Filesystem.php
@@ -2,15 +2,15 @@
 namespace Viserio\Filesystem;
 
 use FilesystemIterator;
+use InvalidArgumentException;
 use League\Flysystem\Util\MimeType;
-use RecursiveDirectoryIterator;
-use RecursiveIteratorIterator;
 use Symfony\Component\Filesystem\Exception\FileNotFoundException as SymfonyFileNotFoundException;
 use Symfony\Component\Filesystem\Exception\IOException as SymfonyIOException;
 use Symfony\Component\Filesystem\Filesystem as SymfonyFilesystem;
+use Symfony\Component\Finder\Finder;
 use Viserio\Contracts\Filesystem\Directorysystem as DirectorysystemContract;
 use Viserio\Contracts\Filesystem\Exception\FileNotFoundException;
-use Viserio\Contracts\Filesystem\Exception\IOException;
+use Viserio\Contracts\Filesystem\Exception\IOException as ViserioIOException;
 use Viserio\Contracts\Filesystem\Filesystem as FilesystemContract;
 use Viserio\Filesystem\Traits\FilesystemExtensionTrait;
 use Viserio\Filesystem\Traits\FilesystemHelperTrait;
@@ -18,9 +18,9 @@ use Viserio\Support\Traits\NormalizePathAndDirectorySeparatorTrait;
 
 class Filesystem extends SymfonyFilesystem implements FilesystemContract, DirectorysystemContract
 {
-    use NormalizePathAndDirectorySeparatorTrait,
-        FilesystemHelperTrait,
-        FilesystemExtensionTrait;
+    use NormalizePathAndDirectorySeparatorTrait;
+    use FilesystemHelperTrait;
+    use FilesystemExtensionTrait;
 
     /**
      * @var array
@@ -86,9 +86,11 @@ class Filesystem extends SymfonyFilesystem implements FilesystemContract, Direct
     {
         $path = $this->normalizeDirectorySeparator($path);
 
-        return file_put_contents($path, $contents, FILE_APPEND);
+        if (!$this->exists($path)) {
+            throw new FileNotFoundException($path);
+        }
 
-        // @TODO throw new FileNotFoundException($path);
+        return file_put_contents($path, $contents, FILE_APPEND);
     }
 
     /**
@@ -111,7 +113,7 @@ class Filesystem extends SymfonyFilesystem implements FilesystemContract, Direct
     public function setVisibility($path, $visibility)
     {
         $path = $this->normalizeDirectorySeparator($path);
-        $visibility = $this->parseVisibility($visibility);
+        $visibility = $this->parseVisibility($path, $visibility);
 
         try {
             $this->chmod($path, $visibility);
@@ -131,12 +133,14 @@ class Filesystem extends SymfonyFilesystem implements FilesystemContract, Direct
         $to   = $this->normalizeDirectorySeparator($targetFile);
 
         try {
-            return parent::copy($from, $to, $override);
+            parent::copy($from, $to, $override);
         } catch (SymfonyFileNotFoundException $exception) {
-            throw FileNotFoundException();
+            throw new FileNotFoundException($exception->getMessage());
         } catch (SymfonyIOException $exception) {
-            throw IOException();
+            throw new ViserioIOException($exception->getMessage());
         }
+
+        return true;
     }
 
     /**
@@ -190,6 +194,8 @@ class Filesystem extends SymfonyFilesystem implements FilesystemContract, Direct
         if (!$this->isFile($path) && !$this->has($path)) {
             throw new FileNotFoundException($path);
         }
+
+        return date('F d Y H:i:s', filemtime($path));
     }
 
     /**
@@ -223,21 +229,10 @@ class Filesystem extends SymfonyFilesystem implements FilesystemContract, Direct
      */
     public function allFiles($directory)
     {
-        $directory = $this->normalizeDirectorySeparator($directory);
-        $recursive = new RecursiveIteratorIterator(
-            new RecursiveDirectoryIterator($directory, RecursiveDirectoryIterator::SKIP_DOTS),
-            RecursiveIteratorIterator::SELF_FIRST,
-            RecursiveIteratorIterator::CATCH_GET_CHILD // Ignore "Permission denied"
-        );
-
         $files = [];
 
-        foreach ($recursive as $file) {
-            if ($file->isDir()) {
-                continue;
-            }
-
-            $files[] = $file->getFilename();
+        foreach (Finder::create()->files()->in($directory) as $dir) {
+            $files[] = $this->normalizeDirectorySeparator($dir->getPathname());
         }
 
         return $files;
@@ -246,12 +241,17 @@ class Filesystem extends SymfonyFilesystem implements FilesystemContract, Direct
     /**
      * {@inheritdoc}
      */
-    public function createDirectory($dirname)
+    public function createDirectory($dirname, array $config = [])
     {
         $dirname = $this->normalizeDirectorySeparator($dirname);
+        $mode    = $this->permissions['dir']['public'];
+
+        if (isset($config['visibility'])) {
+            $mode = $this->permissions['dir'][$config['visibility']];
+        }
 
         try {
-            $this->mkdir($dirname);
+            $this->mkdir($dirname, $mode);
         } catch (SymfonyIOException $exception) {
             return false;
         }
@@ -264,15 +264,13 @@ class Filesystem extends SymfonyFilesystem implements FilesystemContract, Direct
      */
     public function directories($directory)
     {
-        $directory = $this->normalizeDirectorySeparator($directory);
+        $directories = [];
 
-        $dirs = [];
-
-        foreach (glob($directory, GLOB_ONLYDIR) as $dir) {
-            $dirs[] = $this->normalizeDirectorySeparator($dir);
+        foreach (Finder::create()->in($directory)->directories()->depth(0) as $dir) {
+            $directories[] = $this->normalizeDirectorySeparator($dir->getPathname());
         }
 
-        return $dirs;
+        return $directories;
     }
 
     /**
@@ -280,22 +278,7 @@ class Filesystem extends SymfonyFilesystem implements FilesystemContract, Direct
      */
     public function allDirectories($directory)
     {
-        $directory = $this->normalizeDirectorySeparator($directory);
-        $recursive = new RecursiveIteratorIterator(
-            new RecursiveDirectoryIterator($directory, RecursiveDirectoryIterator::SKIP_DOTS),
-            RecursiveIteratorIterator::SELF_FIRST,
-            RecursiveIteratorIterator::CATCH_GET_CHILD // Ignore "Permission denied"
-        );
-
-        $dirs = [];
-
-        foreach ($recursive as $dir) {
-            if ($dir->isDir()) {
-                $dirs[] = $this->normalizeDirectorySeparator($dir->getRealpath());
-            }
-        }
-
-        return $dirs;
+        return iterator_to_array(Finder::create()->directories()->in($directory), false);
     }
 
     /**
@@ -309,11 +292,7 @@ class Filesystem extends SymfonyFilesystem implements FilesystemContract, Direct
             return false;
         }
 
-        try {
-            $this->remove($dirname);
-        } catch (SymfonyIOException $exception) {
-            return false;
-        }
+        $this->remove($dirname);
 
         return true;
     }
@@ -329,13 +308,31 @@ class Filesystem extends SymfonyFilesystem implements FilesystemContract, Direct
             return false;
         }
 
-        $directories = new FilesystemIterator($dirname);
+        $items = new FilesystemIterator($dirname);
 
-        foreach ($directories as $dirname) {
-            @rmdir($dirname);
+        foreach ($items as $item) {
+            if ($item->isDir() && !$item->isLink()) {
+                $this->cleanDirectory($item->getPathname());
+            } else {
+                try {
+                    $this->remove($item->getPathname());
+                } catch (SymfonyIOException $exception) {
+                    return false;
+                }
+            }
         }
 
         return true;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function isDirectory($dirname)
+    {
+        $dirname = $this->normalizeDirectorySeparator($dirname);
+
+        return is_dir($dirname);
     }
 
     /**
@@ -346,7 +343,7 @@ class Filesystem extends SymfonyFilesystem implements FilesystemContract, Direct
      *
      * @throws \InvalidArgumentException
      *
-     * @return int
+     * @return int|null
      */
     private function parseVisibility($path, $visibility)
     {
